[update] finish training

+++ src/data/components/latent_kitti_dataset.py
@@ -14,6 +14,6 @@ class LatentVectorDataset(Dataset):
         gt = np.load(os.path.join(self.root_dir, f"{idx}_gt.npy"))
         rot = np.load(os.path.join(self.root_dir, f"{idx}_rot.npy"))
         w = np.load(os.path.join(self.root_dir, f"{idx}_w.npy"))
-        return (torch.from_numpy(latent_vector).to(torch.float), torch.from_numpy(rot), torch.from_numpy(w)), torch.from_numpy(gt).to(torch.float).squeeze()
+        return (torch.from_numpy(latent_vector).to(torch.float32), torch.from_numpy(rot).to(torch.float32), torch.from_numpy(w).to(torch.float32)), torch.from_numpy(gt).to(torch.float32).squeeze()
